+++ code/ProjectFiles_motorTest/include/blinky_AO.h
@@ -61,7 +61,7 @@ enum blinky_Signals {
 
 typedef struct{
     Event super;
-    int16_t angle;       
+    int32_t angle;       
 }UI_AO_ANGLE_PL;
 
 /* AO Class Data -------------------------------------------------------------*/
